Rename: 독립적 UI 컴포넌트로 활용되는 파일 분리 - 독립적 UI 형태의 컴포넌트 파일 위치 변경 -> Widgets - 공용적으로 관리되는 UI 형태의 컴포넌트 파일 위치 유지 - 스타일 시트 같은 위치의 파일로 이동 - Routing Main 컨테이너 이동 - router 파일 삭제 및 BrowserRouter 활용
