--- src/app/outline/outline.module.css
@@ -1,40 +1,0 @@
-.container {
-  width: 100%;
-  height: 100%;
-  position: absolute;
-  top: 0px;
-  left: 0px;
-  pointer-events: none;
-}
-
-.top {
-  width: 100%;
-  padding: 24px 24px 0px 24px;
-  position: absolute;
-  top: 0px;
-  left: 0px;
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-}
-
-.logo {
-  font-size: 20px;
-  line-height: 20px;
-  position: absolute;
-  top: 24px;
-  left: 24px;
-  font-family: "Roboto", sans-serif;
-  font-weight: bold;
-}
-
-.bottom {
-  width: 100%;
-  padding: 0px 24px 24px 24px;
-  position: absolute;
-  bottom: 0px;
-  left: 0px;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
