--- src/page/index.ts
@@ -1,5 +1,0 @@
-import Main from "./main";
-import Screen from "./screen";
-import Mobile from "./mobile";
-
-export { Main, Screen, Mobile };
