+++ src/shared/components/index.tsx
@@ -1,8 +1,3 @@
 export { default as RealTime } from "./realtime";
-export { default as Navigation } from "./navigation";
-export { default as LockScreen } from "./lock-screen";
-export { default as Applications } from "./applications";
-export { default as Dock } from "./dock";
-export { default as MobileGuide } from "./mobile-guide";
 export { default as Loading } from "./loading";
 export { default as Window } from "./window";
