--- src/shared/router/router.tsx
@@ -1,17 +1,0 @@
-import { createBrowserRouter } from "react-router-dom";
-
-import App from "@/app/app";
-import Screen from "@/page/screen";
-
-const router = createBrowserRouter([
-  {
-    path: "/",
-    element: <App />,
-  },
-  {
-    path: "/screen",
-    element: <Screen />,
-  },
-]);
-
-export default router;
