+++ src/widgets/applications/style/applications.module.css
