+++ src/widgets/doc/style/doc.module.css
