+++ src/widgets/folder-app/style/folder.css
