+++ src/widgets/lock-screen/style/lock-screen.module.css
