+++ src/widgets/memo-app/style/memo.css
