+++ src/widgets/navigation/style/navigation.module.css
