+++ src/widgets/render-apps/style/style.css
