+++ src/widgets/terminal-app/index.tsx
@@ -1,9 +1,21 @@
-import "./style.css";
+import "./style/terminal.css";
 
-import TerminalContainer from "./ui/terminal-container";
+import { Window } from "@/shared/components";
+import TerminalSubNavigation from "./ui/terminal-sub-navigation";
+import TerminalReleaseInfo from "./ui/terminal-release-info";
 
 const Terminal = () => {
-  return <TerminalContainer />;
+  return (
+    <Window name="terminal">
+      <Window.Navigation>
+        <Window.Buttons />
+        <TerminalSubNavigation />
+      </Window.Navigation>
+      <Window.Body>
+        <TerminalReleaseInfo />
+      </Window.Body>
+    </Window>
+  );
 };
 
 export default Terminal;
