+++ src/widgets/terminal-app/style/terminal.css
@@ -84,7 +84,7 @@
 }
 
 .terminal-sub-nav-dir {
-  font-size: 20px;
+  font-size: 18px;
   line-height: 18px;
   color: white;
 }
